PIC16 update

+++ targets/Microchip/PIC16_generic/PIC16_core/api/pin_fast/pin_fast.c.v
@@ -11,18 +11,18 @@ module pin_fast
 
 #include "pin_fast.c"
 
-fn C.HIGH_FAST(port_id int, pin_id int)
-fn C.LOW_FAST(port_id int, pin_id int)
-fn C.READ_FAST(port_id int, pin_id int) u8
-fn C.SETUP_FAST(port_id int, pin_id int, mode u8)
-fn C.TOGGLE_FAST(port_id int, pin_id int)
-fn C.WRITE_FAST(port_id int, pin_id int, value u8)
+fn C.HIGH_FAST(port_id u8, pin_id u8)
+fn C.LOW_FAST(port_id u8, pin_id u8)
+fn C.READ_FAST(port_id u8, pin_id u8) u8
+fn C.SETUP_FAST(port_id u8, pin_id u8, mode u8)
+fn C.TOGGLE_FAST(port_id u8, pin_id u8)
+fn C.WRITE_FAST(port_id u8, pin_id u8, value u8)
 
 // high function puts a logic `1` to a pin faster than the regular high function.
 // `port_id` is the port id, for example `pin_fast.port_a` and
 // `pin_id` is the pin id, for example `pin_fast.b3`.
 @[as_macro]
-pub fn high(port_id int, pin_id int) {
+pub fn high(port_id u8, pin_id u8) {
 	C.HIGH_FAST(port_id, pin_id)
 }
 
@@ -30,7 +30,7 @@ pub fn high(port_id int, pin_id int) {
 // `port_id` is the port id, for example `pin_fast.port_a` and
 // `pin_id` is the pin id, for example `pin_fast.b3`v.
 @[as_macro]
-pub fn low(port_id int, pin_id int) {
+pub fn low(port_id u8, pin_id u8) {
 	C.LOW_FAST(port_id, pin_id)
 }
 
@@ -38,15 +38,16 @@ pub fn low(port_id int, pin_id int) {
 // `port_id` is the port id, for example `pin_fast.port_a` and
 // `pin_id` is the pin id, for example `pin_fast.b3`
 @[as_macro]
-pub fn read(port_id int, pin_id int) u8 {
+pub fn read(port_id u8, pin_id u8) u8 {
 	return C.READ_FAST(port_id, pin_id)
 }
 
 // setup configures the mode of a pin faster than the regular setup function.
-// `port_id` is the port id, for example `pin_fast.port_a` and
+// `port_id` is the port id, for example `pin_fast.port_a`,
 // `pin_id` is the pin id, for example `pin_fast.b3`
+// `mode` is the pin mode, `pin.input` or `pin.output`.
 @[as_macro]
-pub fn setup(port_id int, pin_id int, mode u8) {
+pub fn setup(port_id u8, pin_id u8, mode u8) {
 	C.SETUP_FAST(port_id, pin_id, mode)
 
 }
@@ -55,7 +56,7 @@ pub fn setup(port_id int, pin_id int, mode u8) {
 // `port_id` is the port id, for example `pin_fast.port_a` and
 // `pin_id` is the pin id, for example `pin_fast.b3`
 @[as_macro]
-pub fn toggle(port_id int, pin_id int) {
+pub fn toggle(port_id u8, pin_id u8) {
 	C.TOGGLE_FAST(port_id, pin_id)
 }
 
@@ -63,6 +64,6 @@ pub fn toggle(port_id int, pin_id int) {
 // `port_id` is the port id, for example `pin_fast.port_a` and
 // `pin_id` is the pin id, for example `pin_fast.b3`
 @[as_macro]
-pub fn write(port_id int, pin_id int, value u8) {
+pub fn write(port_id u8, pin_id u8, value u8) {
 	C.WRITE_FAST(port_id, pin_id, value)
 }
